+++ src/main/java/biomesoplenty/common/item/ItemBOPPlant.java
@@ -68,7 +68,7 @@ public class ItemBOPPlant extends ItemBOPBlock {
                         BlockPos blockpos1 = blockpos.up();
                         IBlockState iblockstate = worldIn.getBlockState(blockpos);
 
-                        if (iblockstate.getMaterial() == Material.WATER && iblockstate.getValue(BlockLiquid.LEVEL).intValue() == 0 && worldIn.isAirBlock(blockpos1))
+                        if (iblockstate.getMaterial() == Material.WATER && ((Integer)iblockstate.getValue(BlockLiquid.LEVEL)).intValue() == 0 && worldIn.isAirBlock(blockpos1))
                         {
                             // special case for handling block placement with reeds
                             net.minecraftforge.common.util.BlockSnapshot blocksnapshot = net.minecraftforge.common.util.BlockSnapshot.getBlockSnapshot(worldIn, blockpos1);
